Moving formatters into addon namespace
